xcpmd: Don't leak FDs
Sometimes, logs show "xcpmd: Failed to read
/sys/class/power_supply/BAT0/current_now".  When this message is
triggered, xcpmd is leaking FDs.  Close them when that happens.  There
are other similar code paths that are also updated.

Leaking the FDs, xcpmd can use up its FD limit and no longer be able to
open the sysfs files.  When that happens, there are no more battery
updates.

An HP laptop reports ENODEV for current_now when on battery power -
those errors spam the log and are indicative of the leak scenario.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>