xcpmd: Send battery changed when the percent changes
xcpmd polls the battery every 4 seconds.  When running on battery,
*something* will have changed in that time.  Even on AC, the values have
typically changes such that xcpmd is sending the DBus notification.

However, the only thing that the UI cares about is the battery
percentage.  So check the percentage and only send it if has changed.
This avoids xcpmd sending a message, but it also removes multiple DBus
queries by the UI for more information.

It's a little bit of a micro optimization, but it is also easy.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>