xcpmd: Suppress ENODEV errors
ACPI batteries will return ENODEV when they don't have current_now data.
This is seen with an HP Zbook when on battery power.  It will spam the
logs with: "Failed to read /sys/class/power_supply/BAT0/current_now".

Look for ENODEV and suppress the log message in that case.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>