[xcpmd] Add lid_state_changed dbus signal handler
- added logic to allow xcpmd to receive lid_state_changed signals over
dbus. once that signal is received, xenstore's lid_state value is then
updated. this allows guest vm's to accurately be aware of lid state
changes.
- note that after OpenXT restarts, xenstore will not contain any lid_state
information until the lid is closed/opened. when qemu tries to retrieve
lid state in this situation, it will simply tell guest vm's that the lid
is open to avoid unwanted sleeping/suspending issues.
- code to handle lid events triggered from acpi was removed, as it is
now deprecated and will not be used.

Signed-off-by: Reed McFadden <mcfaddenr@ainfosec.com>
Signed-off-by: Chris Rogers <crogers122@gmail.com>